[REF] l10n_br_account: clean inherited form fields
domain, groups and context where added to analytic, account_id and
product_id fields in the account.move.line form view.
It came from v12 as it. May be it was required on v12,
but on v14 it seems useless as these domain and groups and context are
already present in the parent view:
account.view_move_form in l10n_br_account/views/account_invoice_view.xml

The form view for Brazil is already very complex, it's important to keep
only what is strictly necessary.